benchmarks: moves dependency scope to test
